+++ x/evmutil/keeper/keeper.go
@@ -184,7 +184,7 @@ func (k Keeper) RemoveBalance(ctx sdk.Context, addr sdk.AccAddress, amt sdkmath.
 	return k.SetBalance(ctx, addr, finalBal)
 }
 
-// SetDeployedCosmosCoinContract stores a single deployed ERC20NemoWrappedCosmosCoin contract address
+// SetDeployedCosmosCoinContract stores a single deployed ERC20FuryWrappedCosmosCoin contract address
 func (k *Keeper) SetDeployedCosmosCoinContract(ctx sdk.Context, cosmosDenom string, contractAddress types.InternalEVMAddress) error {
 	if err := sdk.ValidateDenom(cosmosDenom); err != nil {
 		return errorsmod.Wrap(types.ErrInvalidCosmosDenom, cosmosDenom)
@@ -203,7 +203,7 @@ func (k *Keeper) SetDeployedCosmosCoinContract(ctx sdk.Context, cosmosDenom stri
 	return nil
 }
 
-// SetDeployedCosmosCoinContract gets a deployed ERC20NemoWrappedCosmosCoin contract address by cosmos denom
+// SetDeployedCosmosCoinContract gets a deployed ERC20FuryWrappedCosmosCoin contract address by cosmos denom
 // Returns the stored address and a bool indicating if it was found or not
 func (k *Keeper) GetDeployedCosmosCoinContract(ctx sdk.Context, cosmosDenom string) (types.InternalEVMAddress, bool) {
 	store := ctx.KVStore(k.storeKey)